runtime-rs: share fail-closed VMM credentials
The QEMU rootless implementation already fails closed when applying
setgroups(), setgid(), and setuid(). Cloud Hypervisor still ignores
those failures and can execute with inherited supplementary groups,
the wrong primary GID, or even the root shim UID.

Move the credential helper and its injected unit tests from the QEMU
module into common hypervisor utilities. Reuse it for QEMU without
changing its behavior, and call it from the Cloud Hypervisor pre-exec
hook.

Retain coverage for ordering, every failure point, and the empty-group
case that clears inherited supplementary groups.
