+++ src/runtime-rs/crates/hypervisor/src/qemu/inner.rs
@@ -17,7 +17,7 @@ use crate::{
 
 use crate::utils::{
     bytes_to_megs, create_dir_all_with_inherit_owner, enter_netns, get_jailer_root, megs_to_bytes,
-    set_groups, uses_native_ccw_bus, vm_cleanup,
+    set_process_credentials, uses_native_ccw_bus, vm_cleanup,
 };
 
 use anyhow::{anyhow, Context, Result};
@@ -30,7 +30,6 @@ use kata_types::{
     capabilities::{Capabilities, CapabilityBits},
     config::KATA_PATH,
 };
-use nix::unistd::{setgid, setuid, Gid, Uid};
 use persist::sandbox_persist::Persist;
 use qapi_qmp::MigrationStatus;
 use std::cmp::Ordering;
@@ -1041,15 +1040,6 @@ fn check_bpf_enabled_with<ReadStatus, LogWarning>(
     }
 }
 
-fn set_process_credentials(user: &RootlessUser) -> Result<()> {
-    set_process_credentials_with(
-        user,
-        set_groups,
-        |gid| setgid(Gid::from_raw(gid)).map_err(anyhow::Error::from),
-        |uid| setuid(Uid::from_raw(uid)).map_err(anyhow::Error::from),
-    )
-}
-
 fn memlock_limit_with_headroom(guest_memory: u64) -> u64 {
     guest_memory.saturating_add(guest_memory / MEMLOCK_HEADROOM_DIVISOR)
 }
@@ -1063,23 +1053,6 @@ fn set_memlock_rlimit(memlock_limit: u64) -> Result<()> {
     if result != 0 {
         return Err(std::io::Error::last_os_error()).context("set RLIMIT_MEMLOCK failed");
     }
-    Ok(())
-}
-
-fn set_process_credentials_with<SetGroups, SetGid, SetUid>(
-    user: &RootlessUser,
-    set_groups_fn: SetGroups,
-    set_gid_fn: SetGid,
-    set_uid_fn: SetUid,
-) -> Result<()>
-where
-    SetGroups: Fn(&[u32]) -> Result<()>,
-    SetGid: Fn(u32) -> Result<()>,
-    SetUid: Fn(u32) -> Result<()>,
-{
-    set_groups_fn(&user.groups).context("setgroups failed")?;
-    set_gid_fn(user.gid).context("setgid failed")?;
-    set_uid_fn(user.uid).context("setuid failed")?;
     Ok(())
 }
 
@@ -1569,118 +1542,5 @@ mod tests {
             }
             None => assert!(warnings.is_empty()),
         }
-    }
-
-    // The Set* prefix mirrors the setgroups/setgid/setuid syscalls these
-    // variants stand for, so keep it despite clippy::enum_variant_names.
-    #[derive(Debug, PartialEq)]
-    #[allow(clippy::enum_variant_names)]
-    enum CredentialOperation {
-        SetGroups(Vec<u32>),
-        SetGid(u32),
-        SetUid(u32),
-    }
-
-    #[derive(Clone, Copy, Debug, PartialEq)]
-    #[allow(clippy::enum_variant_names)]
-    enum CredentialStep {
-        SetGroups,
-        SetGid,
-        SetUid,
-    }
-
-    fn rootless_user() -> RootlessUser {
-        RootlessUser {
-            uid: 1001,
-            gid: 1002,
-            groups: vec![1003, 1004],
-            user_name: "kata-test".to_string(),
-        }
-    }
-
-    #[rstest]
-    #[case::with_supplementary_groups(vec![1003, 1004])]
-    #[case::without_supplementary_groups(Vec::new())]
-    fn test_set_process_credentials_order(#[case] groups: Vec<u32>) {
-        let operations = RefCell::new(Vec::new());
-        let mut user = rootless_user();
-        user.groups = groups.clone();
-
-        set_process_credentials_with(
-            &user,
-            |groups| {
-                operations
-                    .borrow_mut()
-                    .push(CredentialOperation::SetGroups(groups.to_vec()));
-                Ok(())
-            },
-            |gid| {
-                operations
-                    .borrow_mut()
-                    .push(CredentialOperation::SetGid(gid));
-                Ok(())
-            },
-            |uid| {
-                operations
-                    .borrow_mut()
-                    .push(CredentialOperation::SetUid(uid));
-                Ok(())
-            },
-        )
-        .unwrap();
-
-        assert_eq!(
-            *operations.borrow(),
-            vec![
-                // Calling setgroups with an empty list clears any inherited
-                // supplementary groups and must not be skipped.
-                CredentialOperation::SetGroups(groups),
-                CredentialOperation::SetGid(1002),
-                CredentialOperation::SetUid(1001),
-            ]
-        );
-    }
-
-    #[rstest]
-    #[case::setgroups(CredentialStep::SetGroups, "setgroups failed", 1)]
-    #[case::setgid(CredentialStep::SetGid, "setgid failed", 2)]
-    #[case::setuid(CredentialStep::SetUid, "setuid failed", 3)]
-    fn test_set_process_credentials_stops_on_error(
-        #[case] failed_step: CredentialStep,
-        #[case] expected_error: &str,
-        #[case] expected_calls: usize,
-    ) {
-        let calls = Cell::new(0);
-        let result = set_process_credentials_with(
-            &rootless_user(),
-            |_| {
-                calls.set(calls.get() + 1);
-                if failed_step == CredentialStep::SetGroups {
-                    Err(anyhow!("injected setgroups failure"))
-                } else {
-                    Ok(())
-                }
-            },
-            |_| {
-                calls.set(calls.get() + 1);
-                if failed_step == CredentialStep::SetGid {
-                    Err(anyhow!("injected setgid failure"))
-                } else {
-                    Ok(())
-                }
-            },
-            |_| {
-                calls.set(calls.get() + 1);
-                if failed_step == CredentialStep::SetUid {
-                    Err(anyhow!("injected setuid failure"))
-                } else {
-                    Ok(())
-                }
-            },
-        );
-
-        let error = result.expect_err("credential failure must abort setup");
-        assert!(format!("{error:#}").contains(expected_error));
-        assert_eq!(calls.get(), expected_calls);
     }
 }
