+++ src/runtime-rs/crates/hypervisor/src/utils.rs
@@ -21,7 +21,7 @@ use std::collections::HashMap;
 
 use anyhow::{anyhow, Context, Result};
 use kata_types::{
-    config::{Hypervisor, KATA_PATH},
+    config::{hypervisor::RootlessUser, Hypervisor, KATA_PATH},
     prefix_with_rootless_dir,
 };
 use lazy_static::lazy_static;
@@ -29,7 +29,7 @@ use nix::{
     fcntl,
     sched::{setns, CloneFlags},
     sys::stat,
-    unistd::{chown, setgroups, Gid, Uid},
+    unistd::{chown, setgid, setgroups, setuid, Gid, Uid},
 };
 use rand::{rng, RngExt};
 use serde::{Deserialize, Serialize};
@@ -118,13 +118,37 @@ pub fn enter_netns(netns_path: &str) -> Result<()> {
     Ok(())
 }
 
-pub fn set_groups(groups: &[u32]) -> Result<()> {
-    let group = groups
-        .iter()
-        .map(|gid| Gid::from_raw(*gid))
-        .collect::<Vec<_>>();
-    // An empty list intentionally clears inherited supplementary groups.
-    setgroups(&group).context("set groups failed")?;
+/// Replace supplementary groups and drop the child process to the configured
+/// primary GID and UID. This must run before exec while the child is privileged.
+pub fn set_process_credentials(user: &RootlessUser) -> Result<()> {
+    set_process_credentials_with(
+        user,
+        |groups| {
+            let groups = groups
+                .iter()
+                .map(|gid| Gid::from_raw(*gid))
+                .collect::<Vec<_>>();
+            setgroups(&groups).map_err(anyhow::Error::from)
+        },
+        |gid| setgid(Gid::from_raw(gid)).map_err(anyhow::Error::from),
+        |uid| setuid(Uid::from_raw(uid)).map_err(anyhow::Error::from),
+    )
+}
+
+fn set_process_credentials_with<SetGroups, SetGid, SetUid>(
+    user: &RootlessUser,
+    set_groups_fn: SetGroups,
+    set_gid_fn: SetGid,
+    set_uid_fn: SetUid,
+) -> Result<()>
+where
+    SetGroups: Fn(&[u32]) -> Result<()>,
+    SetGid: Fn(u32) -> Result<()>,
+    SetUid: Fn(u32) -> Result<()>,
+{
+    set_groups_fn(&user.groups).context("setgroups failed")?;
+    set_gid_fn(user.gid).context("setgid failed")?;
+    set_uid_fn(user.uid).context("setuid failed")?;
 
     Ok(())
 }
@@ -520,6 +544,9 @@ pub(crate) fn get_vcpu_tids(proc_path: &str, prefix: &str) -> Result<HashMap<u32
 
 #[cfg(test)]
 mod tests {
+    use anyhow::anyhow;
+    use kata_types::config::hypervisor::RootlessUser;
+    use std::cell::{Cell, RefCell};
     use std::fs;
     use std::os::unix::fs::MetadataExt;
     use std::os::unix::fs::PermissionsExt;
@@ -529,6 +556,7 @@ mod tests {
     use nix::unistd::geteuid;
     use nix::unistd::Gid;
     use nix::unistd::Uid;
+    use rstest::rstest;
     use tempfile::Builder;
     use tempfile::TempDir;
 
@@ -537,8 +565,122 @@ mod tests {
 
     use super::create_fds;
     use super::remove_dir_all_if_exists;
+    use super::set_process_credentials_with;
     use super::vmm_user_runtime_dir;
     use super::SocketAddress;
+
+    // The Set* prefix mirrors the setgroups/setgid/setuid syscalls these
+    // variants stand for, so keep it despite clippy::enum_variant_names.
+    #[derive(Debug, PartialEq)]
+    #[allow(clippy::enum_variant_names)]
+    enum CredentialOperation {
+        SetGroups(Vec<u32>),
+        SetGid(u32),
+        SetUid(u32),
+    }
+
+    #[derive(Clone, Copy, Debug, PartialEq)]
+    #[allow(clippy::enum_variant_names)]
+    enum CredentialStep {
+        SetGroups,
+        SetGid,
+        SetUid,
+    }
+
+    fn rootless_user() -> RootlessUser {
+        RootlessUser {
+            uid: 1001,
+            gid: 1002,
+            groups: vec![1003, 1004],
+            user_name: "kata-test".to_string(),
+        }
+    }
+
+    #[rstest]
+    #[case::with_supplementary_groups(vec![1003, 1004])]
+    #[case::without_supplementary_groups(Vec::new())]
+    fn test_set_process_credentials_order(#[case] groups: Vec<u32>) {
+        let operations = RefCell::new(Vec::new());
+        let mut user = rootless_user();
+        user.groups = groups.clone();
+
+        set_process_credentials_with(
+            &user,
+            |groups| {
+                operations
+                    .borrow_mut()
+                    .push(CredentialOperation::SetGroups(groups.to_vec()));
+                Ok(())
+            },
+            |gid| {
+                operations
+                    .borrow_mut()
+                    .push(CredentialOperation::SetGid(gid));
+                Ok(())
+            },
+            |uid| {
+                operations
+                    .borrow_mut()
+                    .push(CredentialOperation::SetUid(uid));
+                Ok(())
+            },
+        )
+        .unwrap();
+
+        assert_eq!(
+            *operations.borrow(),
+            vec![
+                // Calling setgroups with an empty list clears any inherited
+                // supplementary groups and must not be skipped.
+                CredentialOperation::SetGroups(groups),
+                CredentialOperation::SetGid(1002),
+                CredentialOperation::SetUid(1001),
+            ]
+        );
+    }
+
+    #[rstest]
+    #[case::setgroups(CredentialStep::SetGroups, "setgroups failed", 1)]
+    #[case::setgid(CredentialStep::SetGid, "setgid failed", 2)]
+    #[case::setuid(CredentialStep::SetUid, "setuid failed", 3)]
+    fn test_set_process_credentials_stops_on_error(
+        #[case] failed_step: CredentialStep,
+        #[case] expected_error: &str,
+        #[case] expected_calls: usize,
+    ) {
+        let calls = Cell::new(0);
+        let result = set_process_credentials_with(
+            &rootless_user(),
+            |_| {
+                calls.set(calls.get() + 1);
+                if failed_step == CredentialStep::SetGroups {
+                    Err(anyhow!("injected setgroups failure"))
+                } else {
+                    Ok(())
+                }
+            },
+            |_| {
+                calls.set(calls.get() + 1);
+                if failed_step == CredentialStep::SetGid {
+                    Err(anyhow!("injected setgid failure"))
+                } else {
+                    Ok(())
+                }
+            },
+            |_| {
+                calls.set(calls.get() + 1);
+                if failed_step == CredentialStep::SetUid {
+                    Err(anyhow!("injected setuid failure"))
+                } else {
+                    Ok(())
+                }
+            },
+        );
+
+        let error = result.expect_err("credential failure must abort setup");
+        assert!(format!("{error:#}").contains(expected_error));
+        assert_eq!(calls.get(), expected_calls);
+    }
 
     #[test]
     fn test_ctreate_fds() {
